requirements_auth.txt
# Basketball Analysis Service - Production Authentication & Payment Requirements
# Updated requirements with OAuth2, JWT, and Stripe integration

# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Authentication & Security
Authlib==1.3.0                 # OAuth2 client library
PyJWT==2.8.0                   # JWT token handling
bcrypt==4.1.2                  # Password hashing
cryptography==41.0.8           # General cryptography support

# Google OAuth2 dependencies
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Payment processing
stripe==7.8.0                  # Stripe payment integration

# Database
psycopg2-binary==2.9.9         # PostgreSQL adapter
redis==5.0.1                   # Redis for caching and sessions

# Computer Vision (existing)
mediapipe==0.10.8
opencv-python==4.8.1.78
numpy==1.25.2
scipy==1.11.4

# Machine Learning (existing)
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4

# Video Processing (existing)
moviepy==1.0.3
imageio==2.33.1
imageio-ffmpeg==0.4.9

# AWS SDK
boto3==1.34.19                 # AWS S3, SES, etc.
botocore==1.34.19

# Email handling
flask-mail==0.9.1
premailer==3.10.0              # CSS inlining for emails

# HTTP and API
requests==2.31.0
urllib3==2.1.0

# Validation and serialization
marshmallow==3.20.2
webargs==8.4.0

# Background task processing
celery==5.3.4                  # Distributed task queue
kombu==5.3.4                   # Message passing

# Production server
gunicorn==21.2.0               # WSGI HTTP Server
eventlet==0.33.3               # Async networking

# Monitoring and logging
prometheus-client==0.19.0      # Metrics collection
structlog==23.2.0              # Structured logging

# Configuration and environment
python-dotenv==1.0.0           # Environment variable loading
pydantic==2.5.2                # Data validation

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.11.0                 # Code formatting
flake8==6.1.0                  # Linting
mypy==1.7.1                    # Type checking

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Rate limiting and security
flask-limiter==3.5.0           # API rate limiting
flask-talisman==1.1.0          # Security headers

# Session management
flask-session==0.5.0

# Timezone handling
pytz==2023.3

# Data processing utilities
python-dateutil==2.8.2
click==8.1.7

# Health checks and monitoring
health-check==1.3.3

# File handling
pillow==10.1.0                 # Image processing
python-magic==0.4.27          # File type detection

# API documentation
flask-restx==1.3.0             # API documentation and validation

# Environment-specific packages
# Development
ipython==8.18.1                # Enhanced Python shell
jupyter==1.0.0                 # Notebook environment

# Production optimizations
gevent==23.9.1                 # Async library for production
